reorganized folders in makefile
