+++ src/prompt/prompt.c
@@ -1,4 +1,4 @@
-#include "../include/minishell.h"
+#include "../../include/minishell.h"
 
 void	check_prompt(char *line)
 {
